==> 01_load_and_clean.ipynb <==
import pandas as pd

### Importing Datasets 

df1 = pd.read_csv(r"C:\Users\jhanvi.kasundra\Downloads\Projects New\Data/df1.csv")   #job title and desc

ml_jobs = pd.read_csv(r"C:\Users\jhanvi.kasundra\Downloads\Projects New\Data/1000_ml_jobs_us.csv")

resumes = pd.read_csv(r"C:\Users\jhanvi.kasundra\Downloads\Projects New\Data/preprocessed_dataset.csv") # cat, text 

print ("df1:", df1.shape, df1.columns)
print("ml_jobs:", ml_jobs.shape, ml_jobs.columns)
print ("resumes:", resumes.shape, resumes.columns)

### Cleaning & Standardizing Jobs into on schema 

jobs_a = df1.rename(columns={"title": "job_title","description":"job_text"})  ##renaming cols
jobs_a["company_name"]=None  ## adding new col
jobs_a["location"]=None   ## adding new col
jobs_a["source"]="df1"  ##Add a new column called source, and write df1 in every row

jobs_b = pd.DataFrame({
    "job_title":ml_jobs["job_title"],
    "job_text":ml_jobs["job_description_text"],
    "company_name":ml_jobs["company_name"],
    "location":ml_jobs["company_address_locality"].fillna("")+", " + ml_jobs["company_address_region"].fillna(""),
    "source":"ml_jobs"
})

jobs = pd.concat([jobs_a[["job_title","job_text","company_name","location","source"]],
                  jobs_b[["job_title","job_text","company_name","location","source"]]], ignore_index=True)

### Clean minimal (remove empty text)

jobs = jobs.dropna(subset=["job_text"])
jobs = jobs[jobs["job_text"].str.strip().astype(bool)]

resumes = resumes.dropna(subset=["text_cv"])
resumes = resumes[resumes["text_cv"].str.strip().astype(bool)]

### Saving clean version of datasets

jobs.to_csv(r"C:\Users\jhanvi.kasundra\Downloads\Projects New\Data/jobs_clean.csv", index=False)
resumes.to_csv(r"C:\Users\jhanvi.kasundra\Downloads\Projects New\Data/resumes_clean.csv", index=False)

print("jobs_clean:", jobs.shape, "resumes_clean:", resumes.shape)


jobs.head()

resumes.head()
